+++ sertop/sertop_async.ml
@@ -38,10 +38,11 @@ let sertop_init ~(fb_out : Sexp.t -> unit) ~iload_path ~require_libs ~debug ~all
   let fb_handler fb = Sertop.Sertop_ser.sexp_of_answer (Feedback (Sertop.Sertop_util.feedback_tr fb)) |> fb_out in
 
   coq_init {
-    fb_handler;
-    ml_load = None;
-    debug;
-    allow_sprop;
+    fb_handler
+  ; ml_load = None
+  ; debug
+  ; allow_sprop
+  ; indices_matter = false
   };
 
   let stm_options = process_stm_flags