+++ sertop/sertop_sexp.mli
@@ -29,6 +29,7 @@ type ser_opts =
 
 ; debug    : bool                (** Enable Coq debug mode                    *)
 ; allow_sprop: bool              (** Allow using the proof irrelevant SProp sort *)
+; indices_matter : bool          (** Indices of indexes contribute to inductive level *)
 ; print0   : bool                (** End every answer with [\0]               *)
 ; lheader  : bool                (** Print lenght header (deprecated)         *)